Refactor primitive and task modules to domain-specific subdirectories
