+++ lib/domains/logo/makeLogoTasks.py
@@ -137,7 +137,7 @@ def parseLogo(s):
 
 def manualLogoTask(name, expression, proto=False, needToTrain=False, supervise = False):
     p = parseLogo(expression)
-    from lib.primitives.logoPrimitives import primitives
+    from lib.domains.logo.logoPrimitives import primitives
     from grammar import Grammar
     g = Grammar.uniform(primitives, continuationType=turtle)
     gp = Grammar.uniform(primitives)
@@ -699,7 +699,7 @@ if __name__ == "__main__":
     montageTasks(rotationalSymmetryDemo(),"rotational")
 
     from grammar import *
-    from lib.primitives.logoPrimitives import *
+    from lib.domains.logo.logoPrimitives import *
 
     g0 = Grammar.uniform(primitives, continuationType=turtle)
     eprint("dreaming into /tmp/dreams_0...")
