+++ lib/domains/misc/napsPrimitives.py
