+++ lib/domains/puddleworld/makePuddleworldTasks.py
@@ -3,8 +3,8 @@ Makes Puddleworld tasks.
 Tasks are (gridworld, text instruction) -> goal coordinate.
 Credit: tasks are taken from: https://github.com/JannerM/spatial-reasoning 
 """
-from puddleworldPrimitives import *
-from lib.tasks.task import *
+from lib.domains.puddleworld.puddleworldPrimitives import *
+from lib.task import *
 from type import *
 
 OBJECT_NAMES = ["NULL", "puddle", "star", "circle", "triangle", "heart", "spade", "diamond", "rock", "tree", "house", "horse"]
