+++ puddleworld.py
@@ -11,8 +11,8 @@ from ec import commandlineArguments
 from grammar import Grammar
 from utilities import eprint, numberOfCPUs
 
-from puddleworldPrimitives import primitives
-from lib.tasks.makePuddleworldTasks import makeLocalTasks, makeGlobalTasks
+from lib.domains.puddleworld.puddleworldPrimitives import primitives
+from lib.domains.puddleworld.makePuddleworldTasks import makeLocalTasks, makeGlobalTasks
 
 
 def puddleworld_options(parser):
