+++ rational.py
@@ -1,12 +1,5 @@
-from ec import explorationCompression, commandlineArguments, Program
-from fragmentGrammar import *
-from grammar import Grammar
-from arithmeticPrimitives import real_addition, real_multiplication, real_power, real_subtraction, real, f0, f1, fpi, real_division
-from listPrimitives import bootstrapTarget
-from lib.tasks.task import DifferentiableTask, squaredErrorLoss, l1loss, Task
-from type import tint, arrow
-from utilities import *
-from program import *
+from ec import explorationCompression, commandlineArguments
+from lib.domains.arithmetic.arithmeticPrimitives import real_addition, real_multiplication, real, real_division
 from recognition import *
 
 import random
